+++ library/rustsbi/CHANGELOG.md
@@ -10,6 +10,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Add the SBI Message Proxy extension trait and dispatch support.
+- Add the SBI Debug Triggers extension trait and dispatch support.
 
 ### Modified